requirements.txt
# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.0


httpx==0.27.0
beautifulsoup4==4.12.3
duckduckgo-search==6.3.5

openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.0
pytest==8.3.0
pytest-asyncio==0.24.0